Add interfaces for the API so that it is cleaner
Add interfaces for the API so that it does not show the methods that shouldn't be accessible by default.
