+++ src/main/java/net/quickwrite/fluent4j/ast/FluentElement.java
@@ -1,10 +1,10 @@
 package net.quickwrite.fluent4j.ast;
 
-import net.quickwrite.fluent4j.FluentBundle;
 import net.quickwrite.fluent4j.ast.placeable.NumberLiteral;
 import net.quickwrite.fluent4j.ast.placeable.SelectExpression;
 import net.quickwrite.fluent4j.util.args.FluentArgs;
 import net.quickwrite.fluent4j.util.args.FluentArgument;
+import net.quickwrite.fluent4j.util.bundle.ResourceFluentBundle;
 
 /**
  * The base interface for the AST.
@@ -23,11 +23,11 @@ public interface FluentElement {
      * does some extra checks so that things like plural rules are
      * getting considered.
      *
-     * @param bundle The base bundle
+     * @param bundle   The base bundle
      * @param selector The other element
      * @return If both are matching
      */
-    boolean matches(final FluentBundle bundle, final FluentArgument selector);
+    boolean matches(final ResourceFluentBundle bundle, final FluentArgument selector);
 
     /**
      * Returns the String value of the object
@@ -42,9 +42,9 @@ public interface FluentElement {
      * the different arguments and the base bundle that
      * it has been called from.
      *
-     * @param bundle The base bundle
+     * @param bundle    The base bundle
      * @param arguments The arguments that are being passed on the scope
      * @return The {@link CharSequence} value of the argument with the parameters
      */
-    CharSequence getResult(final FluentBundle bundle, final FluentArgs arguments);
+    CharSequence getResult(final ResourceFluentBundle bundle, final FluentArgs arguments);
 }
