+++ src/main/java/net/quickwrite/fluent4j/ast/placeable/AttributeReference.java
@@ -1,6 +1,5 @@
 package net.quickwrite.fluent4j.ast.placeable;
 
-import net.quickwrite.fluent4j.FluentBundle;
 import net.quickwrite.fluent4j.ast.FluentAttribute;
 import net.quickwrite.fluent4j.ast.FluentElement;
 import net.quickwrite.fluent4j.ast.FluentMessage;
@@ -10,6 +9,8 @@ import net.quickwrite.fluent4j.ast.placeable.base.FluentSelectable;
 import net.quickwrite.fluent4j.util.StringSlice;
 import net.quickwrite.fluent4j.util.args.FluentArgs;
 import net.quickwrite.fluent4j.util.args.FluentArgument;
+import net.quickwrite.fluent4j.util.bundle.FluentBundle;
+import net.quickwrite.fluent4j.util.bundle.ResourceFluentBundle;
 
 import java.util.List;
 
@@ -32,7 +33,7 @@ public class AttributeReference implements FluentPlaceable, FluentArgumentResult
     }
 
     @Override
-    public boolean matches(final FluentBundle bundle, final FluentArgument selector) {
+    public boolean matches(final ResourceFluentBundle bundle, final FluentArgument selector) {
         return false;
     }
 
@@ -42,7 +43,7 @@ public class AttributeReference implements FluentPlaceable, FluentArgumentResult
     }
 
     @Override
-    public CharSequence getResult(final FluentBundle bundle, final FluentArgs arguments) {
+    public CharSequence getResult(final ResourceFluentBundle bundle, final FluentArgs arguments) {
         final FluentMessage fluentMessage = this.getMessage(bundle, reference.stringValue());
         if (fluentMessage == null) {
             return getErrorString();
@@ -58,7 +59,7 @@ public class AttributeReference implements FluentPlaceable, FluentArgumentResult
     }
 
     @Override
-    public FluentArgument getArgumentResult(final FluentBundle bundle, final FluentArgs arguments) {
+    public FluentArgument getArgumentResult(final ResourceFluentBundle bundle, final FluentArgs arguments) {
         final FluentAttribute attribute = this.getMessage(bundle, reference.stringValue())
                 .getAttribute(this.attributeIdentifier);
         if (attribute == null) {
@@ -75,7 +76,7 @@ public class AttributeReference implements FluentPlaceable, FluentArgumentResult
         return (FluentArgument) elementList.get(0);
     }
 
-    protected FluentMessage getMessage(final FluentBundle bundle, final String key) {
+    protected FluentMessage getMessage(final ResourceFluentBundle bundle, final String key) {
         return bundle.getMessage(key);
     }
 
@@ -97,7 +98,7 @@ public class AttributeReference implements FluentPlaceable, FluentArgumentResult
         }
 
         @Override
-        protected FluentMessage getMessage(final FluentBundle bundle, final String key) {
+        protected FluentMessage getMessage(final ResourceFluentBundle bundle, final String key) {
             return bundle.getTerm(key);
         }
