+++ src/main/java/net/quickwrite/fluent4j/ast/placeable/NumberLiteral.java
@@ -6,12 +6,12 @@ import com.ibm.icu.number.NumberFormatter;
 import com.ibm.icu.text.NumberFormat;
 import com.ibm.icu.text.PluralRules;
 import com.ibm.icu.util.ULocale;
-import net.quickwrite.fluent4j.FluentBundle;
 import net.quickwrite.fluent4j.ast.placeable.base.FluentPlaceable;
 import net.quickwrite.fluent4j.ast.placeable.base.FluentSelectable;
 import net.quickwrite.fluent4j.util.StringSlice;
 import net.quickwrite.fluent4j.util.args.FluentArgs;
 import net.quickwrite.fluent4j.util.args.FluentArgument;
+import net.quickwrite.fluent4j.util.bundle.ResourceFluentBundle;
 
 import java.math.BigDecimal;
 
@@ -50,7 +50,7 @@ public class NumberLiteral implements FluentPlaceable, FluentSelectable {
     }
 
     @Override
-    public CharSequence getResult(final FluentBundle bundle, final FluentArgs arguments) {
+    public CharSequence getResult(final ResourceFluentBundle bundle, final FluentArgs arguments) {
         return NumberFormat.getInstance(bundle.getLocale()).format(number);
     }
 
@@ -69,7 +69,7 @@ public class NumberLiteral implements FluentPlaceable, FluentSelectable {
     }
 
     @Override
-    public boolean matches(final FluentBundle bundle, final FluentArgument selector) {
+    public boolean matches(final ResourceFluentBundle bundle, final FluentArgument selector) {
         if (selector instanceof NumberLiteral) {
             return matches(((NumberLiteral) selector).valueOf());
         }
