+++ src/main/java/net/quickwrite/fluent4j/ast/placeable/VariableReference.java
@@ -1,12 +1,12 @@
 package net.quickwrite.fluent4j.ast.placeable;
 
-import net.quickwrite.fluent4j.FluentBundle;
 import net.quickwrite.fluent4j.ast.placeable.base.FluentArgumentResult;
 import net.quickwrite.fluent4j.ast.placeable.base.FluentPlaceable;
 import net.quickwrite.fluent4j.ast.placeable.base.FluentSelectable;
 import net.quickwrite.fluent4j.util.StringSlice;
 import net.quickwrite.fluent4j.util.args.FluentArgs;
 import net.quickwrite.fluent4j.util.args.FluentArgument;
+import net.quickwrite.fluent4j.util.bundle.ResourceFluentBundle;
 
 /**
  * Variables are pieces of data received from the app.
@@ -41,7 +41,7 @@ public class VariableReference implements FluentPlaceable, FluentSelectable, Flu
     }
 
     @Override
-    public boolean matches(final FluentBundle bundle, final FluentArgument selector) {
+    public boolean matches(final ResourceFluentBundle bundle, final FluentArgument selector) {
         return selector.stringValue().equals(content);
     }
 
@@ -51,7 +51,7 @@ public class VariableReference implements FluentPlaceable, FluentSelectable, Flu
     }
 
     @Override
-    public FluentArgument getArgumentResult(FluentBundle bundle, final FluentArgs arguments) {
+    public FluentArgument getArgumentResult(ResourceFluentBundle bundle, final FluentArgs arguments) {
         final FluentArgument argument = arguments.getNamed(content);
 
         if (argument == null) {
@@ -62,7 +62,7 @@ public class VariableReference implements FluentPlaceable, FluentSelectable, Flu
     }
 
     @Override
-    public CharSequence getResult(final FluentBundle bundle, final FluentArgs arguments) {
+    public CharSequence getResult(final ResourceFluentBundle bundle, final FluentArgs arguments) {
         final FluentArgument argument = arguments.getNamed(content);
 
         if (argument == null) {
