+++ src/main/java/net/quickwrite/fluent4j/ast/placeable/base/FluentFunction.java
@@ -1,10 +1,11 @@
 package net.quickwrite.fluent4j.ast.placeable.base;
 
-import net.quickwrite.fluent4j.FluentBundle;
+import net.quickwrite.fluent4j.util.args.FluentArgs;
+import net.quickwrite.fluent4j.util.bundle.ResourceFluentBundle;
 import net.quickwrite.fluent4j.exception.FluentParseException;
 import net.quickwrite.fluent4j.util.StringSlice;
 import net.quickwrite.fluent4j.util.StringSliceUtil;
-import net.quickwrite.fluent4j.util.args.FluentArgs;
+import net.quickwrite.fluent4j.util.args.ResourceFluentArguments;
 import net.quickwrite.fluent4j.util.args.FluentArgument;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.commons.lang3.tuple.Pair;
@@ -33,7 +34,7 @@ public abstract class FluentFunction implements FluentPlaceable, FluentArgumentR
     }
 
     private FluentArgs getArguments(final StringSlice content) {
-        FluentArgs arguments = new FluentArgs();
+        FluentArgs arguments = new ResourceFluentArguments();
 
         while (!content.isBigger()) {
             StringSliceUtil.skipWhitespaceAndNL(content);
@@ -85,7 +86,7 @@ public abstract class FluentFunction implements FluentPlaceable, FluentArgumentR
      * @param arguments The arguments in this scope
      * @return The sanitized arguments of the function
      */
-    protected FluentArgs getArguments(final FluentBundle bundle, final FluentArgs arguments) {
+    protected FluentArgs getArguments(final ResourceFluentBundle bundle, final FluentArgs arguments) {
         this.arguments.sanitize(bundle, arguments);
         return this.arguments;
     }
@@ -97,7 +98,7 @@ public abstract class FluentFunction implements FluentPlaceable, FluentArgumentR
      * @param arguments The arguments that are passed into this function
      * @return The resulting {@link FluentArgument} that has been created
      */
-    public abstract FluentArgument getArgumentResult(final FluentBundle bundle, final FluentArgs arguments);
+    public abstract FluentArgument getArgumentResult(final ResourceFluentBundle bundle, final FluentArgs arguments);
 
     /**
      * Checks if this FluentFunction and the selector are the same.
@@ -107,7 +108,7 @@ public abstract class FluentFunction implements FluentPlaceable, FluentArgumentR
      * @return If they are the same object
      */
     @Override
-    public boolean matches(final FluentBundle bundle, final FluentArgument selector) {
+    public boolean matches(final ResourceFluentBundle bundle, final FluentArgument selector) {
         return this.equals(selector);
     }
