+++ src/test/java/net/quickwrite/fluent4j/TestAstral.java
@@ -1,5 +1,7 @@
 package net.quickwrite.fluent4j;
 
+import net.quickwrite.fluent4j.util.bundle.FluentBundle;
+import net.quickwrite.fluent4j.util.bundle.ResourceFluentBundle;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.jupiter.api.Assertions;
